requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and LangGraph
langchain==0.1.0
langchain-google-genai==0.0.5
langgraph==0.0.20

# Google AI
google-generativeai==0.3.0

# Supabase
supabase==2.0.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0